install dependencies separately, put all falcon apps into one python service
