+++ container/sagemaker/nginx.conf.template
@@ -32,6 +32,7 @@ http {
     subrequest_output_buffer_size 100m;
 
     set $tfs_version %TFS_VERSION%;
+    set $default_tfs_model %TFS_DEFAULT_MODEL_NAME%;
 
     location /tfs {
         rewrite ^/tfs/(.*) /$1  break;
