+++ docker/1.11/Dockerfile.eia
@@ -15,7 +15,8 @@ RUN \
 
 # cython, falcon, gunicorn
 RUN \
-    pip3 install cython falcon gunicorn gevent requests tensorflow-serving-api
+    pip3 install cython falcon gunicorn gevent requests grpcio protobuf tensorflow && \
+    pip3 install --no-dependencies tensorflow-serving-api
 
 COPY ./ /
 RUN mv amazonei_tensorflow_model_server /usr/bin/tensorflow_model_server && \
